+++ src/main/java/com/portfolio/BackEnd/repository/ProyectoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProyectoRepository extends JpaRepository<Proyecto, Long>{
+public interface ProyectoRepository extends JpaRepository<Proyecto, Integer>{
     
 }
